Improve hint path
The previous method used line numbers, so editing code prior to the class would have changed the hint path.
The new method uses the index in Members.
Moreover, since Roslyn has been updated to allow the use of / and \, the hint path is based on the full path of the original source.